# optimized lua storage benchmark