before : 
```txt
Starting storage benchmark ...
storage: iterations=5000 average: 614us total: 3069000us

Total Runtime: 3069000us
```


after : 

```txt
Starting storage benchmark ...
storage: iterations=5000 average: 303us total: 1514000us

Total Runtime: 1514000us
```

ran on LuaJIT HEAD