@smarr 2020-10-31
Thanks for putting the effort into this.

And after sleeping over it, and looking through all the other implementations of the benchmark, I am slowly arriving at the view that this is likely not the right way to go. The real solution should be something along the lines of https://research.google/pubs/pub43823/

With respect to using FFI, I believe this is also not what these benchmarks should do. One can of course derive a variation that focuses on FFI performance. Though, this should then be very explicit about it.

To add to my explanation from earlier: different from other benchmark games, in the Are We Fast Yet benchmarks, I am not looking for the fastest way one could write these programs. Instead, I am looking to use a restricted subset of each language, that is as comparable as possible across languages, which in turn allows us to identify optimization potential in our runtime systems.