@ExtReMLapin 2020-10-31
Smarr, i had a discussion with @GitSparTV, I told him I find the existence of this benchmark and the used implementation quite surprising.

Following this PDF : http://software.rochus-keller.ch/are-we-fast-yet_lua_results_2020-10-12.pdf

We find that most of the places where LuaJIT under-performs are the places where closures are used.

* Richards
* Havlak


`Storage` under-performed for no obvious reason, this is why I decided to fix the performances issues.


Comparing javascript, a technology built around the web and [asynchronous work/privacy] to Lua, a technology clearly not meant to be used in this context, seems wrong to me.

However, when we move to a pure-arithmetics benchmarks like mandelbrot, LuaJIT destroys V8 JIT compiler with no surprise.



>Instead, I am looking to use a restricted subset of each language, that is as comparable as possible across languages, which in turn allows us to identify optimization potential in our runtime systems.

But is it the right way ?


Some languages will have features that other will never have, for example, in Lua, you can fill the table on creature with `{n = 4, functioncall(), functioncall(), functioncall(), functioncall())`

Alright, you can't do it with javascript (or can you ?) but that's part of the languages war, there are some stuff that can and cannot be optimized automatically and the dev is supposed to make the correct steps to let the VM understand what is going or what is going to happen.

This is the point of static typing, as a developer, you do an extra step but it's for the greater good.


Shouldn't the implementation follow strictly pseudocode implementation like ? : 


`FILL AN ARRAY WITH 9 RANDOM INTEGERS`

`RANDOMIZE THE ARRAY ORDER`

Then do it in your language, the best way you can. I don't feel like benchmarking is just about the language, it's also about how the developer uses it.

Obviously, let's not fall into those *"beating C with 80 lines of haskell"* memes.