@eregon 2020-10-31
As a 3rd point of view, I tend to agree with @smarr.
It seems the idiomatic way to create an array and fill it element by element in Lua is https://www.lua.org/pil/11.1.html,
assuming the length is not constant, or that duplicating the code for each element in the source is not wanted (this is the case here).
That said, I don't know much about Lua, so maybe I missed something?

Given this kind of code is likely to appear frequently in user code, it seems the job of the VM to optimize it.
For instance, if the example from the docs:
```lua
    a = {}    -- new array
    for i=1, 1000 do
      a[i] = 0
    end
```
is slow on LuaJIT, then that seems the real problem.