@smarr 2020-10-31
My impression is we simply have different goals. I want to compare the runtime systems and compilers, not the languages.

The reasoning is detailed here: https://stefan-marr.de/papers/dls-marr-et-al-cross-language-compiler-benchmarking-are-we-fast-yet/

> Comparing javascript, a technology built around the web and [asynchronous work/privacy] to Lua, a technology clearly not meant to be used in this context, seems wrong to me.

Why would it be wrong? The languages benefit from very similar runtime systems and optimizations.
Again, for me, this is not about the languages. I am interested in implementation technology.

> But is it the right way ?

It is for the goal I have in mind, at least based on my current understanding (see for the above write up).

> you can fill the table on creature with

The syntax for instance for JavaScript is only minimally different, it works for objects and for arrays. Though, this is not the point. The point is that I want to stress specific parts of a runtime system, and ideally the same parts, as far as that is possibly between languages.

>  there are some stuff that can and cannot be optimized automatically

Doing research on this myself, the border is much further out than what one may think. There are a lot of know tricks that many implementations simply don't do. And yes, one reason is they often come with tradeoffs.

> Shouldn't the implementation follow strictly pseudocode implementation like ? :
> Then do it in your language, the best way you can. I don't feel like benchmarking is just about the language, it's also about how the developer uses it.

One can design benchmarks that way, and that's what the [Computer Language Benchmarks Game](https://benchmarksgame-team.pages.debian.net/benchmarksgame/index.html) does. Are We Fast Yet has very different goals. And I believe those to be useful, too.