@krono 2020-10-31
I'd like to chime in.
In my view these benchmarks are not meant to answer the question “How fast can you possibly go?”.
But rather—with restrictions—“If performance was for free, how would you write it? And then, how fast is it?”.
I see it as an _idiomatic baseline_ and not merely C-in-disguise.