@ExtReMLapin 2020-10-31
> For instance, if the example from the docs:
> 
> ```lua
>     a = {}    -- new array
>     for i=1, 1000 do
>       a[i] = 0
>     end
> ```
> 
> is slow on LuaJIT, then that seems the real problem.

Here i'm going to blame the developer, if you're writing this code when knowing you'll go with luajit, you're doing something wrong