@smarr 2020-10-31
Thanks again for the contribution. At this point, I believe it's not something that is in alignment with the goals of my version of AWFY. As such, I'll close the PR.

That being said, I am sure there are other possible improvements, and I am happy to consider changes, as long as they fit within the project's vision.

Thank you